Setting the font shear in a BPicture didn't work, because the value was treated as uint32 in some places, when it's a float.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@22214 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/private/interface/PictureDataWriter.h
@@ -45,7 +45,7 @@ public:
 	status_t WriteSetFontRotation(const float &rotation);
 	status_t WriteSetFontEncoding(const int32 &encoding);
 	status_t WriteSetFontFlags(const int32 &flags);
-	status_t WriteSetFontShear(const int32 &shear);
+	status_t WriteSetFontShear(const float &shear);
 	status_t WriteSetFontFace(const int32 &face);
 
 	status_t WriteStrokeLine(const BPoint &start, const BPoint &end);
